refactor(kodo): flatten 2-layer stack (PR 1b)
Eliminate kodo's architectural inconsistency. Every other backend has
only backends/<name>/. Kodo uniquely had backends/kodo/ wrapping
adapters/kodo/, predating the backend layer. After this PR:
backends/kodo/ owns both the contract layer and the raw runner layer.

Structural changes (no behavior change)
- Moved adapters/kodo/adapter.py → backends/kodo/runner.py
- Deleted src/operations_center/adapters/kodo/ entirely
- Updated 9 import sites + 1 docstring reference

After this PR
- backends/kodo/runner.py owns the raw subprocess + kodo-binary layer
- backends/kodo/invoke.py uses runner.KodoAdapter directly (no
  cross-package import)
- _run_subprocess is now isolated in one file — clean Phase 3
  extraction target for ExecutorRuntime

2407 unit tests pass.
