+++ .console/log.md
@@ -5,6 +5,8 @@ _Not a task tracker — that's backlog.md. Keep entries concise and dated._
 
 ## Stop Points
 
+- PR 1b — kodo two-layer flatten (2026-05-06, on `oc-kodo-flatten-adapter`): Eliminated kodo's architectural inconsistency. Every other backend has only `backends/<name>/`; kodo uniquely had a 2-layer stack (`backends/kodo/` over `adapters/kodo/`). Moved `adapters/kodo/adapter.py` → `backends/kodo/runner.py` (new file, raw subprocess + command-build layer; KodoAdapter, KodoRunResult, _run_subprocess, _get_kodo_version). Deleted `src/operations_center/adapters/kodo/` entirely (was 219 LOC across 2 files). Bulk-rewrote 9 import sites (3 src/ + 6 tests) + 1 docstring reference. Pure structural relocation — no behavior change. 2407 unit tests pass (unchanged from PR 1a). Phase 3 extraction target (`_run_subprocess` → ExecutorRuntime) is now isolated in one file.
+
 - PR 1a — kodo dead-fallback cut (2026-05-06, on `oc-kodo-cut-dead-fallback`): Removed proven-dead Kodo fallback machinery from `adapters/kodo/` and `backends/kodo/`. Empirical audit (read-only, 338 real runs) showed: zero `quota_exhausted: true` records, zero `rate_limited: true` records, zero codex orchestrator usage in commands, zero matches for any of the four removed signal-pattern lists in real artifacts. **Cut**: 4 signal-list constants (`_CODEX_QUOTA_SIGNALS`, `_ORCHESTRATOR_RATE_LIMIT_SIGNALS`, `_HARD_QUOTA_EXHAUSTED_SIGNALS`, `_SONNET_EXHAUSTED_SIGNALS`), 2 fallback team configs (`_CLAUDE_FALLBACK_TEAM`, `_OPUS_HAIKU_FALLBACK_TEAM` — adapter copies; binder's parallel copies stay since they're the canonical RuntimeBinding-driven path), 6 detector methods (`_is_codex_quota_error`, `is_orchestrator_rate_limited`, `_is_sonnet_exhausted`, `is_quota_exhausted`, `_run_with_team`, `_run_with_claude_fallback`), 1 module-level shim (`_is_quota_exhausted_result`). Simplified `KodoAdapter.run()` to: build_command → _run_subprocess → return. Dropped now-dead `rate_limited`/`quota_exhausted` fields from `KodoRunCapture` and `is_rate_limited`/`is_quota_exhausted` fields from `KodoFailureInfo`. Dropped duplicate `_RATE_LIMITED_SIGNALS`/`_QUOTA_EXHAUSTED_SIGNALS` from `backends/kodo/normalize.py`. Removed the same fields from artifact JSON emission in `backends/kodo/adapter.py:240`. Pre-existing `.kodo/team.json` double-writer race eliminated as side effect — binder is now sole writer (preserves RuntimeBinding integrity). 2407 unit tests pass (was 2408 before; -1 from removed `_is_quota_exhausted_result` shim test). Net: -324 / +18 lines. Real production "out of extra usage" gap NOT addressed in this PR — filed as G-005 follow-up.
 
 - Source-registry hard-cut (2026-05-06, on `main`): Removed all back-compat surfaces from the SR migration. **OC**: deleted `src/operations_center/entrypoints/upstream/` (deprecation shim) + `operations-center-upstream` entry from pyproject.toml — users now invoke `source-registry` directly with `--registry registry/source_registry.yaml [--patches registry/patches]`. Deleted `UnknownExecutorError` from `execution/binding.py` (never raised). Updated `UnknownBackendError` docstring to reflect actual semantics (envelope.backend missing or not in catalog). Updated `docs/architecture/execution_target.md`. **SR companion**: removed `InstallKind.PYTHON_TOOL` legacy alias (was a seed-compat shim aliasing to `cli_tool`); test renamed accordingly. SR 66 tests pass (1 unchanged), OC 2408 tests pass. Cross-repo audit clean — no other repo (SwitchBoard/WorkStation/OperatorConsole/Custodian/VideoFoundry/CxRP/RxP/ExecutorRuntime) referenced any removed surface.
