--- src/operations_center/adapters/kodo/__init__.py
@@ -1,5 +1,0 @@
-# SPDX-License-Identifier: AGPL-3.0-or-later
-# Copyright (C) 2026 Velascat
-from operations_center.adapters.kodo.adapter import KodoAdapter, KodoRunResult
-
-__all__ = ["KodoAdapter", "KodoRunResult"]
